Split routes
