+++ packages/utils/src/zod.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod'
+
+export const stringifiedBooleanSchema = z
+  .enum(['true', 'false'])
+  .default('false')
+  .transform(v => v === 'true')
